made passcode screens code modular.
