+++ src/screens/passcodenew.js
@@ -1,84 +1,26 @@
-import React, { useState } from "react";
+import React from "react";
 import {
   ThemeProvider,
-  createTheme,
   CssBaseline,
   Container,
   Typography,
-  TextField,
-  Button,
   Box,
   Link,
 } from "@mui/material";
-import { motion } from "framer-motion";
-import { useNavigate } from "react-router-dom";
-import { encryptSeedPhrase } from "../security/encryption";
-import { saveEncryptedSeedToLocalStorage } from "../security/storage";
-import { useLocation } from "react-router-dom";
+import { darkTheme } from "../theme/darkTheme";
+import { PasscodeForm } from "../components/passcodeForm";
+import { AnimatedButton } from "../components/animatedButton";
+import { usePasscodeSetup } from "../hooks/usePasscodeSetup";
 
-
-const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-    primary: {
-      main: "#ff6b00",
-    },
-    background: {
-      default: "#121212",
-      paper: "#1D1D1D",
-    },
-    text: {
-      primary: "#FFFFFF",
-      secondary: "#AAAAAA",
-    },
-  },
-  typography: {
-    fontFamily: "Roboto, sans-serif",
-    h5: {
-      fontWeight: 500,
-    },
-    button: {
-      textTransform: "none",
-    },
-  },
-});
-
-const buttonVariants = {
-  hover: {
-    scale: 1.1,
-    transition: {
-      duration: 0.3,
-      yoyo: Infinity,
-    },
-  },
-};
-
-function Passcodenew() {
-  const [newPassword, setNewPassword] = useState("");
-  const [repeatPassword, setRepeatPassword] = useState("");
-  const navigate = useNavigate();
-  const location = useLocation();
-const seedPhrase = location.state?.seedPhrase;
-
-  const handleBack = () => {
-    navigate(-1);
-  };
-
-  const handleContinue = () => {
-    if (newPassword === repeatPassword && newPassword.length > 0) {
-      // Encrypt seed phrase and save to local storage
-      console.log("original seedphrase:" + seedPhrase);
-      const encryptedSeed = encryptSeedPhrase(seedPhrase, newPassword);
-      console.log("Original seed phrase:", seedPhrase);
-      console.log("Encrypted seed phrase:", encryptedSeed);
-      saveEncryptedSeedToLocalStorage(encryptedSeed);
-
-      console.log("Seed phrase encrypted and stored successfully!");
-      navigate("/portfolio", { state: { password: newPassword } }); // navigate to the next screen
-    } else {
-      alert("Passwords do not match or are empty.");
-    }
-  };
+function PasscodeNew() {
+  const {
+    newPassword,
+    setNewPassword,
+    repeatPassword,
+    setRepeatPassword,
+    handleBack,
+    handleContinue,
+  } = usePasscodeSetup();
 
   return (
     <ThemeProvider theme={darkTheme}>
@@ -93,50 +35,24 @@ const seedPhrase = location.state?.seedPhrase;
           Set a passcode for your wallet
         </Typography>
 
-        <Box
-          component="div"
-          sx={{
-            backgroundColor: darkTheme.palette.background.paper,
-            padding: "24px",
-            borderRadius: "8px",
-            marginBottom: "24px",
-          }}
-        >
-          <TextField
-            variant="outlined"
-            fullWidth
-            margin="normal"
-            label="New Password"
-            type="password"
-            value={newPassword}
-            onChange={(e) => setNewPassword(e.target.value)}
-          />
-          <TextField
-            variant="outlined"
-            fullWidth
-            margin="normal"
-            label="Repeat Password"
-            type="password"
-            value={repeatPassword}
-            onChange={(e) => setRepeatPassword(e.target.value)}
-          />
-        </Box>
+        <PasscodeForm
+          newPassword={newPassword}
+          repeatPassword={repeatPassword}
+          setNewPassword={setNewPassword}
+          setRepeatPassword={setRepeatPassword}
+        />
 
         <Box display="flex" justifyContent="space-between">
-          <motion.div variants={buttonVariants} whileHover="hover">
-            <Button variant="text" color="primary" onClick={handleBack}>
-              BACK
-            </Button>
-          </motion.div>
-          <motion.div variants={buttonVariants} whileHover="hover">
-            <Button
-              variant="contained"
-              color="primary"
-              onClick={handleContinue}
-            >
-              CONTINUE
-            </Button>
-          </motion.div>
+          <AnimatedButton variant="text" color="primary" onClick={handleBack}>
+            BACK
+          </AnimatedButton>
+          <AnimatedButton
+            variant="contained"
+            color="primary"
+            onClick={handleContinue}
+          >
+            CONTINUE
+          </AnimatedButton>
         </Box>
 
         <Typography
@@ -155,4 +71,4 @@ const seedPhrase = location.state?.seedPhrase;
   );
 }
 
-export default Passcodenew;
+export default PasscodeNew;
