+++ src/theme/darkTheme.js
@@ -1,0 +1,27 @@
+import { createTheme } from "@mui/material";
+
+export const darkTheme = createTheme({
+  palette: {
+    mode: "dark",
+    primary: {
+      main: "#ff6b00",
+    },
+    background: {
+      default: "#121212",
+      paper: "#1D1D1D",
+    },
+    text: {
+      primary: "#FFFFFF",
+      secondary: "#AAAAAA",
+    },
+  },
+  typography: {
+    fontFamily: "Roboto, sans-serif",
+    h5: {
+      fontWeight: 500,
+    },
+    button: {
+      textTransform: "none",
+    },
+  },
+});
